@@ -4,13 +4,11 @@ module Travis
   module Live
     module Middleware
       class Logging
-        def call(worker, message, queue, &block)
-          time = Benchmark.ms do
-            yield
-          end
+        def call(_worker, message, queue, &block)
+          time = Benchmark.ms(&block)
         ensure
           uuid, _, _, payload, params = *message['args']
-          data = Hash.new.tap do |data|
+          data = {}.tap do |data|
             data['type'] = queue
             if payload['build']
               data['build'] = payload['build']['id']
@@ -18,14 +16,12 @@ module Travis
               data['build'] = payload['build_id']
             end
 
-            if payload['repository']
-              data['repo'] = payload['repository']['slug']
-            end
+            data['repo'] = payload['repository']['slug'] if payload['repository']
 
             data['event'] = params['event'] if params['event']
             data['uuid'] = uuid
             data['job'] = payload['id'] if params['event'] && params['event'] =~ /^job/
-            data['time'] = "%.3f" % (time/1000) if time
+            data['time'] = format('%.3f', (time / 1000)) if time
             data['jid'] = message['jid']
           end
           log(data)
