@@ -9,7 +9,7 @@ module Travis
 
         sidekiq_options dead: false
 
-        def perform(uuid, target, method, payload, params)
+        def perform(_uuid, _target, _method, payload, params)
           Travis::Live::Pusher::Task.new(payload, params).run
         end
       end
