@@ -1,5 +1,5 @@
 module Travis
   module Live
-    VERSION = "0.0.1"
+    VERSION = '0.0.1'
   end
 end
